+++ packages/thirdweb/src/wallets/wallet-connect/consumer/index.ts
@@ -1,12 +1,13 @@
 import { SignClient } from "@walletconnect/sign-client";
 import { hexToNumber } from "viem";
-import { hexToBigInt } from "../../utils/encoding/hex.js";
-import type { Prettify } from "../../utils/type-utils.js";
-import type { Account, Wallet } from "../interfaces/wallet.js";
-import { getDefaultAppMetadata } from "../utils/defaultDappMetadata.js";
-import { DEFAULT_PROJECT_ID } from "./constants.js";
+import { hexToBigInt } from "../../../utils/encoding/hex.js";
+import type { Prettify } from "../../../utils/type-utils.js";
+import type { Account, Wallet } from "../../../wallets/interfaces/wallet.js";
+import { getDefaultAppMetadata } from "../../utils/defaultDappMetadata.js";
+import { DEFAULT_PROJECT_ID } from "../constants.js";
 import type {
   WalletConnectConfig,
+  WalletConnectRawTransactionRequestParams,
   WalletConnectSession,
   WalletConnectSessionEvent,
   WalletConnectSessionProposalEvent,
@@ -14,7 +15,8 @@ import type {
   WalletConnectSignRequestPrams,
   WalletConnectSignTypedDataRequestParams,
   WalletConnectTransactionRequestParams,
-} from "./types.js";
+} from "../types.js";
+import { onSessionProposal } from "./session-proposal.js";
 
 export type WalletConnectClient = Awaited<ReturnType<typeof SignClient.init>>;
 
@@ -93,30 +95,6 @@ export function getActiveWalletConnectSession(
   wallet: Wallet,
 ): WalletConnectSession | undefined {
   return walletConnectSessions.get(wallet);
-}
-
-/**
- * @internal
- */
-export async function onSessionProposal(options: {
-  wallet: Wallet;
-  walletConnectClient: WalletConnectClient;
-  event: WalletConnectSessionProposalEvent;
-}) {
-  const { wallet, walletConnectClient, event } = options;
-  const account = wallet.getAccount();
-  if (!account) {
-    throw new Error("[WalletConnect] No account connected to provided wallet");
-  }
-
-  await disconnectExistingSession({ wallet, walletConnectClient });
-  const session = await acceptSessionProposal({
-    account,
-    walletConnectClient,
-    sessionProposal: event.params,
-  });
-  console.log("Session", session);
-  walletConnectSessions.set(wallet, session);
 }
 
 /**
@@ -203,7 +181,6 @@ export async function fulfillRequest(options: {
     }
     case "eth_sendTransaction": {
       const chainId = Number.parseInt(rawChainId?.split(":")[1] ?? "0"); // chainId is of the form "eip155:1234"
-      console.log("CHAIN ID", chainId);
       if (!chainId) {
         throw new Error(
           `[WalletConnect] Invalid chainId ${rawChainId}, request chainId should have the format 'eip155:1'`,
@@ -236,6 +213,30 @@ export async function fulfillRequest(options: {
       result = txHash;
       break;
     }
+    case "eth_sendRawTransaction": {
+      if (!account.sendRawTransaction) {
+        throw new Error(
+          "[WalletConnect] The current account does not support sending raw transactions",
+        );
+      }
+
+      const chainId = Number.parseInt(rawChainId?.split(":")[1] ?? "0"); // chainId is of the form "eip155:1234"
+      if (!chainId) {
+        throw new Error(
+          `[WalletConnect] Invalid chainId ${rawChainId}, request chainId should have the format 'eip155:1'`,
+        );
+      }
+
+      const rawTransaction = (
+        request.params as WalletConnectRawTransactionRequestParams
+      )[0];
+      const txHash = await account.sendRawTransaction({
+        rawTransaction,
+        chainId,
+      });
+      result = txHash;
+      break;
+    }
     default:
       throw new Error(
         `[WalletConnect] Unsupported request method: ${request.method}`,
@@ -250,76 +251,4 @@ export async function fulfillRequest(options: {
       result,
     },
   });
-}
-
-async function disconnectExistingSession({
-  wallet,
-  walletConnectClient,
-}: { wallet: Wallet; walletConnectClient: WalletConnectClient }) {
-  if (hasActiveWalletConnectSession(wallet)) {
-    const existingSession = getActiveWalletConnectSession(
-      wallet,
-    ) as WalletConnectSession;
-
-    await walletConnectClient.disconnect({
-      topic: existingSession.topic,
-      reason: { code: 6000, message: "New session" }, // Code 6000 is user disconnected
-    });
-
-    walletConnectSessions.delete(wallet);
-  }
-}
-
-async function acceptSessionProposal({
-  account,
-  walletConnectClient,
-  sessionProposal,
-}: {
-  account: Account;
-  walletConnectClient: WalletConnectClient;
-  sessionProposal: WalletConnectSessionProposalEvent["params"];
-}): Promise<WalletConnectSession> {
-  if (!sessionProposal.requiredNamespaces.eip155) {
-    throw new Error(
-      "[WalletConnect] No EIP155 namespace found in Wallet Connect session proposal",
-    );
-  }
-
-  if (!sessionProposal.requiredNamespaces.eip155.chains) {
-    throw new Error(
-      "[WalletConnect] No chains found in EIP155 Wallet Connect session proposal namespace",
-    );
-  }
-
-  const namespaces = {
-    chains: [
-      ...sessionProposal.requiredNamespaces.eip155.chains.map(
-        (chain: string) => `${chain}:${account.address}`,
-      ),
-      ...(sessionProposal.optionalNamespaces?.eip155?.chains?.map(
-        (chain: string) => `${chain}:${account.address}`,
-      ) ?? []),
-    ],
-    methods: [
-      ...sessionProposal.requiredNamespaces.eip155.methods,
-      ...(sessionProposal.optionalNamespaces?.eip155?.methods ?? []),
-    ],
-    events: [
-      ...sessionProposal.requiredNamespaces.eip155.events,
-      ...(sessionProposal.optionalNamespaces?.eip155?.events ?? []),
-    ],
-  };
-  const approval = await walletConnectClient.approve({
-    id: sessionProposal.id,
-    namespaces: {
-      eip155: {
-        accounts: namespaces.chains,
-        methods: namespaces.methods,
-        events: namespaces.events,
-      },
-    },
-  });
-
-  const session = await approval.acknowledged();
-  return session;
 }
